fix: change KvCache from Array4 to Array5 to match ONNX model output shape
- Update position_ids type from f32 to i64 to match ONNX model input
- Fix decoder_init and decoder_step to handle 5D KV cache tensors
- Update session manager trait and mock implementations
- Fix integration test model path

+++ src-tauri/stt-core/src/lib.rs
@@ -71,13 +71,13 @@ pub trait SessionManager: Send + Sync {
     fn run_decoder_init(
         &mut self,
         input_embeds: &ndarray::Array3<f32>,
-        position_ids: &ndarray::Array2<f32>,
+        position_ids: &ndarray::Array2<i64>,
     ) -> Result<(u32, KvCache)>;
 
     fn run_decoder_step(
         &mut self,
         input_embeds: &ndarray::Array3<f32>,
-        position_ids: &ndarray::Array2<f32>,
+        position_ids: &ndarray::Array2<i64>,
         past_keys: &ndarray::Array4<f32>,
         past_values: &ndarray::Array4<f32>,
     ) -> Result<(u32, KvCache)>;
